move the TestTagCopying to ExifToolHelper test script fix TestTagCopying to use the TMP_DIR instead of creating (and leaving) an arbitrary test directory in the TEMP location of the system fix bug in ExifToolHelper where passing in common_args=None will cause a crash
